requirements.txt
streamlit
pandas
numpy
plotly
PyYAML
SQLAlchemy
pyodbc
bcrypt
